+++ src/Validation/dtos/validation.isvalidmongoid.dto.ts
@@ -1,0 +1,3 @@
+import { ValidationGeneralDto } from "./validation.general.dto";
+
+export interface IValidationIsValidMongoIdDto extends Omit<ValidationGeneralDto, 'params' | 'customFunction' > {}